Return stats in O(1)

+++ lib/src/main/kotlin/com/test/stream/Event.kt
@@ -1,3 +1,3 @@
 package com.test.stream
 
-data class Event(val occurredOn: Long, val x: Double, val y: Double)
+data class Event(val occurredOn: Long, val x: Double, val y: Long)
